Create WorkingGas.py
this spits out if the robot is within 3 meters of the cylinders. The frame needs to be in odom to see the robot move with teleop. if you want to see the markers add marker topic and /gas_markers and change frame to map. just type in map into fixed frame